Just use onfocusin as a detection test for focusin support
Due to various restrictions on what dojo/on can depend on,
and a desire to resolve this issue so that a release can be issued,
simply check for a matching event key on the has-test element.
This stops focusin/focusout from being used in WebKit and Blink
in lieu of event capturing, which is not ideal, but is not any
worse than what was happening before.

Refs #17599.